Update actuators to use WPILib typed units API
- Motor methods now use Current and Voltage types instead of doubles
- Servo methods now use Angle and Current types
- Updated all implementations (TalonFX, NEO, Minion, servos)
- Updated Javadoc examples to use typed units syntax
- Added backwards-compatible double overloads where appropriate

+++ src/main/java/com/adambots/lib/actuators/config/CurrentLimitConfig.java
@@ -45,7 +45,7 @@ public record CurrentLimitConfig(
      * @param limitRpmThreshold RPM threshold below which stall limit applies
      * @return A new CurrentLimitConfig instance
      */
-    public static CurrentLimitConfig fromAmps(int stallLimitAmps, int freeLimitAmps, int limitRpmThreshold) {
+    public static CurrentLimitConfig fromAmps(double stallLimitAmps, double freeLimitAmps, int limitRpmThreshold) {
         return new CurrentLimitConfig(
             Amps.of(stallLimitAmps),
             Amps.of(freeLimitAmps),
